+++ saleor/discount/tests/test_discounts.py
@@ -6,6 +6,7 @@ import pytest
 from django.utils import timezone
 from prices import Money, TaxedMoney
 
+from ...checkout.fetch import CheckoutLineInfo
 from ...discount.interface import VariantPromotionRuleInfo
 from .. import DiscountValueType, RewardValueType, VoucherType
 from ..models import (
@@ -16,6 +17,7 @@ from ..models import (
     VoucherCustomer,
 )
 from ..utils import (
+    _get_the_cheapest_line,
     activate_voucher_code,
     add_voucher_usage_by_customer,
     deactivate_voucher_code,
@@ -637,3 +639,33 @@ def test_is_order_level_voucher_another_type(voucher_type, voucher):
 
     # then
     assert result is False
+
+
+def test_get_the_cheapest_line_no_lines_provided():
+    # when
+    line_info = _get_the_cheapest_line(None)
+    # then
+    assert line_info is None
+
+
+def test_get_the_cheapest_line(checkout_with_items, channel_USD):
+    # given
+    lines = [
+        CheckoutLineInfo(
+            line=line,
+            channel_listing=line.variant.channel_listings.first(),
+            collections=[],
+            product=line.variant.product,
+            variant=line.variant,
+            discounts=list(line.discounts.all()),
+            rules_info=[],
+            product_type=line.variant.product.product_type,
+            channel=channel_USD,
+            voucher=None,
+        )
+        for line in checkout_with_items.lines.all()
+    ]
+    # when
+    line_info = _get_the_cheapest_line(lines)
+    # then
+    assert line_info == lines[0]
